Waits and displays URL for hidden services
Tor will automatically create the "hostname" file. Let's wait for that
file to be created (on first run it will take a few seconds, thereafter
it'll always be there), then read its contents and print them to the
screen for easier documentation.

Reloads tor service if new HS was created

On first run, the new hidden service won't actually be created, along
with its 'hostname' file, until the service has been bounced. Let's
register the service directory creation task and bounce the tor service
immediately if it changed, rather than waiting for the role to trigger
handlers. If we don't run the change immediately, the `wait_for` task
will block indefinitely.

Not using `meta: flush_handlers` because that would trigger all stacked
handlers in the current play to run. A more selective approach is
prudent.